+++ generic/generate/main.go
@@ -9,30 +9,102 @@ import (
 )
 
 var typeLetters = []string{"A", "B", "C", "D", "E", "F", "G", "H"}
+var numbers = []string{"zero", "one", "two", "three", "four", "five", "six", "seven", "eight"}
+var indices = []string{"zeroth", "first", "second", "third", "fourth", "fifth", "sixth", "seventh", "eighth"}
 
 type query struct {
 	Index       int
+	NumberStr   string
 	Types       string
 	TypesFull   string
 	TypesReturn string
 	ReturnAll   string
 	Include     string
+	Components  string
+	Arguments   string
 }
 type getter struct {
-	Query query
-	ID    int
-	Index int
-	Type  string
+	Query    query
+	ID       int
+	Index    int
+	IndexStr string
+	Type     string
 }
 
 func main() {
-	fmt.Println("Generating code")
+	generateQueries()
+	generateMutates()
+}
+
+func generateMutates() {
+	fmt.Println("Generating mutates")
+
+	maxIndex := len(typeLetters)
+	text := bytes.Buffer{}
+
+	header, err := template.ParseFiles("./generate/mutate_header.gotxt")
+	if err != nil {
+		panic(err)
+	}
+	err = header.Execute(&text, struct{}{})
+	if err != nil {
+		panic(err)
+	}
+
+	mutate, err := template.ParseFiles("./generate/mutate.gotxt")
+	if err != nil {
+		panic(err)
+	}
+
+	for i := 1; i <= maxIndex; i++ {
+		returnAll := ""
+		components := ""
+		arguments := ""
+
+		types := "[" + strings.Join(typeLetters[:i], ", ") + "]"
+		returnTypes := "*" + strings.Join(typeLetters[:i], ", *")
+		fullTypes := "[" + strings.Join(typeLetters[:i], " any, ") + " any]"
+		include := "[]ecs.ID{ecs.ComponentID[" + strings.Join(typeLetters[:i], "](w), ecs.ComponentID[") + "](w)}"
+		for j := 0; j < i; j++ {
+			returnAll += fmt.Sprintf("(*%s)(m.world.Get(entity, m.ids[%d]))", typeLetters[j], j)
+			arguments += fmt.Sprintf("%s *%s", strings.ToLower(typeLetters[j]), typeLetters[j])
+			if j < i-1 {
+				returnAll += ", "
+				arguments += ", "
+			}
+			components += fmt.Sprintf("ecs.Component{ID: m.ids[%d], Component: %s},\n", j, strings.ToLower(typeLetters[j]))
+		}
+
+		data := query{
+			Index:       i,
+			NumberStr:   numbers[i],
+			Types:       types,
+			TypesReturn: returnTypes,
+			TypesFull:   fullTypes,
+			ReturnAll:   returnAll,
+			Include:     include,
+			Components:  components,
+			Arguments:   arguments,
+		}
+		err = mutate.Execute(&text, data)
+		if err != nil {
+			panic(err)
+		}
+	}
+
+	if err := os.WriteFile("mutate_generated.go", text.Bytes(), 0666); err != nil {
+		panic(err)
+	}
+}
+
+func generateQueries() {
+	fmt.Println("Generating queries")
 
 	maxIndex := len(typeLetters)
 
 	text := bytes.Buffer{}
 
-	header, err := template.ParseFiles("./generate/header.gotxt")
+	header, err := template.ParseFiles("./generate/query_header.gotxt")
 	if err != nil {
 		panic(err)
 	}
@@ -64,16 +136,19 @@ func main() {
 			types = "[" + strings.Join(typeLetters[:i], ", ") + "]"
 			returnTypes = "*" + strings.Join(typeLetters[:i], ", *")
 			fullTypes = "[" + strings.Join(typeLetters[:i], " any, ") + " any]"
-			include = "include: []reflect.Type{typeOf[" + strings.Join(typeLetters[:i], "](), typeOf[") + "]()},"
+			include = "[]reflect.Type{typeOf[" + strings.Join(typeLetters[:i], "](), typeOf[") + "]()}"
 			for j := 0; j < i; j++ {
 				returnAll += fmt.Sprintf("(*%s)(q.Query.Get(q.ids[%d]))", typeLetters[j], j)
 				if j < i-1 {
 					returnAll += ", "
 				}
 			}
+		} else {
+			include = "[]reflect.Type{}"
 		}
 		data := query{
 			Index:       i,
+			NumberStr:   numbers[i],
 			Types:       types,
 			TypesReturn: returnTypes,
 			TypesFull:   fullTypes,
@@ -95,10 +170,11 @@ func main() {
 
 		for j := 0; j < i; j++ {
 			getterData := getter{
-				Query: data,
-				ID:    j + 1,
-				Index: j,
-				Type:  typeLetters[j],
+				Query:    data,
+				ID:       j + 1,
+				Index:    j,
+				IndexStr: indices[j+1],
+				Type:     typeLetters[j],
 			}
 			err = queryGet.Execute(&text, getterData)
 			if err != nil {